refactor: QueryParameter 클래스로 쿼리 스트링 처리
